Added support for Windows XP-Theming, based on BCGControlBar-library 6.2.
As a developer you will need to download that new version of BCG, because I have moved all the BCG related stuff from the CVS tree out of license reasons. Take a look into the new version of the "HowToBuild.txt" to learn where to download the necessary files for BCG 6.2.