Cleaned up the code responsible for saving text documents
The old code was difficult to understand and error prone. Specifically, the
separated move/delete operation performed on temporary and backup files didn't
guarantee that the content of original document will not be lost. For instance,
if writing to a temporary file and deleting the original file succeeds, but the
following replacement of the original file by the created temporary for some
reason fails, the content of the original file will be lost.

This changeset removes the usage of unsafe path operations used to generate
temporary file names and replaces DeleteFile/MoveFile by MoveFileEx. In case
moving a temporary file fails, the original document will remain intact.  Also,
the temporary file will be not be deleted allowing the user to recover the
document content that should have been saved.